Abstracted Subsystem, and implemented SingleExecutionSubsystems and Continuous Subsystems.
